Migrate runtime tests to ci specs
