+++ app/helper/ensure.js
@@ -149,76 +149,4 @@ function ensure(param, is, strictly, recursive) {
   return { and: ensure };
 }
 
-function tests() {
-  var assert = require("assert");
-
-  var model = {
-    a: [{ b: "string", c: "number" }],
-    d: { g: { h: { i: [{ j: "string", k: "number" }] } } },
-  };
-
-  var test = {
-    a: [
-      { b: "x", c: 1 },
-      { b: "y", c: 2 },
-      { b: "z", c: 3 },
-    ],
-    d: {
-      g: {
-        h: {
-          i: [
-            { j: "", k: 0 },
-            { j: "", k: 2 },
-          ],
-        },
-      },
-    },
-  };
-
-  assert(ensure(test, model));
-
-  var model2 = { a: { b: { c: "string" } } };
-
-  var test2 = { a: { b: { c: 1 } } };
-
-  assert.throws(function () {
-    ensure(test2, model2);
-  }, Error);
-
-  var model3 = { a: [{ b: "number", d: [{ e: "string", f: "number" }] }] };
-
-  var test3 = {
-    a: [
-      {
-        b: 1,
-        d: [
-          { e: "1", f: 2 },
-          { e: "2", f: 3 },
-          { e: "3", f: 4 },
-        ],
-      },
-      {
-        b: 2,
-        d: [
-          { e: "3", f: 5 },
-          { e: "4", f: 6 },
-          { e: "5", f: 7 },
-        ],
-      },
-    ],
-  };
-
-  assert(ensure(test3, model3));
-
-  var model4 = { a: "string", b: "number", c: "boolean", d: { e: "string" } };
-
-  var test4 = { a: "defghi", b: 12345555, c: "       " };
-
-  assert.throws(function () {
-    ensure(test4, model4);
-  }, Error);
-}
-
-// tests();
-
 module.exports = ensure;
