+++ app/helper/formJSON.js
@@ -2,27 +2,6 @@ var ensure = require("./ensure");
 var type = require("./type");
 var arrayify = require("./arrayify");
 
-// if (fields.menu) {
-
-//   updates.menu = [];
-
-//   for (var i in fields) {
-
-//     if (i.slice(0, 6) === 'title_') {
-
-//       var id = i.slice(6),
-//           link = {
-//             id: id,
-//             url: fields['url_' + id],
-//             label: fields[i]
-//           };
-
-//       if (link.url && link.label)
-//         updates.menu.push(link);
-//     }
-//   }
-// }
-
 // how to implement some rudimentary type coercion?
 // would be nice to coerce strings -> numbs
 // and strings -> bools
@@ -159,90 +138,6 @@ function parseBool(string) {
     default:
       return Boolean(string);
   }
-}
-
-function unitTests() {
-  var assert = require("assert");
-
-  var field1 = {
-    "foo.baz": 1,
-    "foo.bar": 2,
-    "foo.bat.cat": 3,
-    "foo.bat.bar": 4,
-  };
-
-  var model1 = {
-    foo: {
-      baz: "number",
-      bar: "number",
-      bat: {
-        cat: "number",
-        bar: "string",
-      },
-    },
-  };
-
-  var result1 = {
-    foo: {
-      baz: 1,
-      bar: 2,
-      bat: {
-        cat: 3,
-        bar: "4",
-      },
-    },
-  };
-
-  // assert.deepEqual(formJSON(field1, model1), result1);
-
-  var field2 = {
-    "foo.bar.1.title": "Hello",
-    "foo.bar.1.url": "Bye",
-    "foo.bar.1.id": "1",
-    "foo.bar.2.title": "OK",
-    "foo.bar.2.url": "GO",
-    "foo.bar.2.id": "2",
-  };
-
-  var model2 = {
-    foo: {
-      bar: [{ id: "string", title: "string", url: "string" }],
-    },
-  };
-
-  var expected2 = {
-    foo: {
-      bar: [
-        { title: "Hello", url: "Bye", id: "1" },
-        { title: "OK", url: "GO", id: "2" },
-      ],
-    },
-  };
-
-  var resulted = {
-    foo: {
-      bar: [
-        {
-          title: "Hello",
-          url: "Bye",
-          id: "1",
-        },
-        {
-          title: "OK",
-          url: "GO",
-          id: "2",
-        },
-      ],
-    },
-  };
-
-  var result2 = formJSON(field2, model2);
-
-  // console.log(JSON.stringify(result2));
-
-  // assert.deepEqual(resulted, expected2);
-
-  assert.deepEqual(result2, expected2);
 }
 
 module.exports = formJSON;
