+++ app/helper/nameFrom.js
@@ -31,32 +31,4 @@ function nameFrom(src) {
   return name || UID(10);
 }
 
-var assert = require("assert");
-
-function is(src, expected) {
-  var output = nameFrom(src);
-
-  try {
-    assert(output === expected);
-  } catch (e) {
-    console.log("INPUT:", src);
-    console.log("OUTPUT:", output);
-    console.log("EXPECTED:", expected);
-  }
-}
-
-is(
-  "/foo0000000000000000000000000000000000000000.txt",
-  "0000000000000000000000000000000.txt"
-);
-
-is("/foo.txt", "foo.txt");
-is("/foo/bar.txt", "bar.txt");
-is("bar.txt", "bar.txt");
-
-is("//google.com/bar.txt", "bar.txt");
-is("https://google.com/fOo-bAr.txt?baz=true&twit=false", "foo-bar.txt");
-
-// is('http://google.com');
-
 module.exports = nameFrom;
